Only calling window.localStorage and window.sessionStorage after safe check

+++ JavaScript/JavaScriptSDK.Tests/CheckinTests/DataLossAnalyzer.tests.ts
@@ -8,7 +8,9 @@ class DataLossAnalyzerTests extends TestClass {
     }
 
     public testInialize() {
-        sessionStorage.clear();
+        if (Microsoft.ApplicationInsights.Util.canUseSessionStorage()) {
+            sessionStorage.clear();
+        }
     }
 
     public registerTests() {
